% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunctions.R
\name{DNH_iPSE}
\alias{DNH_iPSE}
\title{Compute the iPSE in natural history of human diseases}
\usage{
DNH_iPSE(
  data,
  zvec = NULL,
  plot_result = TRUE,
  iPSEweights = TRUE,
  bootstrap = FALSE,
  boot_times = 1000,
  num_of_cores = 1,
  for_bootstrap = FALSE,
  push_warnings = TRUE,
  plot_unit = 1,
  save_plot = FALSE,
  plot_name = NULL,
  folder_name = NULL,
  match_ylim = FALSE,
  PM = FALSE,
  timer = TRUE,
  Med_TTEM = FALSE,
  sensitivity_analysis = FALSE,
  sensitivity_analysis_match_ylim = FALSE
)
}
\arguments{
\item{data}{a data.frame where columns must be in the order of T1, d1, T2, d2, T3, d3, Z}
}
\value{
A list of self-described result
}
\description{
Compute the iPSE in natural history of human diseases
}
